[DEV-98] refactor: 폴더 구조 컨벤션 준수
