+++ src/components/containers/tab-sidemenu/logo-component.tsx
